+++ DefenderAPI/changelog.md
@@ -1,8 +1,9 @@
 ﻿# Changelog
 
-## ???
+## 0.9.8 (2024-05-22)
 
 + New: MDCA Connection and Commands
++ Upd: Connect-DefenderAPI - Added support for direct Key Vault authentication
 + Upd: Migrated authentication to use external module "EntraAuth"
 
 ## 0.8.5 (2024-04-05)